Simplify skipping create Kafka topics when needed

+++ examples/mqtt.kafka.broker/docker/compose/setup.sh
@@ -13,14 +13,16 @@ NAMESPACE=zilla-mqtt-kafka-broker
 if [[ -z $(docker-compose -p $NAMESPACE ps -q zilla) ]]; then
   docker-compose -p $NAMESPACE up -d
 
-  # Create the mqtt topics in Kafka
-  docker run --rm bitnami/kafka:3.2 bash -c "
-  echo 'Creating topics for $KAFKA_HOST:$KAFKA_PORT'
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-messages
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-devices --config cleanup.policy=compact
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-retained --config cleanup.policy=compact
-  /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-sessions --config cleanup.policy=compact
-  "
+  if [[ "$KAFKA_SKIP_CREATE_TOPICS" != "yes" ]]; then
+    # Create the mqtt topics in Kafka
+    docker run --rm bitnami/kafka:3.2 bash -c "
+    echo 'Creating topics for $KAFKA_HOST:$KAFKA_PORT'
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-messages
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-devices --config cleanup.policy=compact
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-retained --config cleanup.policy=compact
+    /opt/bitnami/kafka/bin/kafka-topics.sh --bootstrap-server $KAFKA_HOST:$KAFKA_PORT --create --if-not-exists --topic mqtt-sessions --config cleanup.policy=compact
+    "
+  fi
 
 else
   docker-compose -p $NAMESPACE restart --no-deps zilla
